Solution clean up
- Update project readme, folder structures, remove project redundancies.
- Pre-refactor and rewritten.
