--- CmdSpaceship.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.168
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CmdSpaceship", "CmdSpaceship.vcxproj", "{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		ClientRelease|x64 = ClientRelease|x64
-		ClientRelease|x86 = ClientRelease|x86
-		ServerRelease|x64 = ServerRelease|x64
-		ServerRelease|x86 = ServerRelease|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientRelease|x64.ActiveCfg = ClientRelease|x64
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientRelease|x64.Build.0 = ClientRelease|x64
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientRelease|x86.ActiveCfg = ClientRelease|Win32
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientRelease|x86.Build.0 = ClientRelease|Win32
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerRelease|x64.ActiveCfg = ServerRelease|x64
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerRelease|x64.Build.0 = ServerRelease|x64
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerRelease|x86.ActiveCfg = ServerRelease|Win32
-		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerRelease|x86.Build.0 = ServerRelease|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {35F3DEE3-E7BE-42A2-87EE-7D48440AA441}
-	EndGlobalSection
-EndGlobal
