+++ CmdSpaceship/CmdSpaceship.sln
@@ -1,0 +1,30 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35931.197
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CmdSpaceship", "CmdSpaceship.vcxproj", "{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		ClientBuild|x64 = ClientBuild|x64
+		ClientBuild|x86 = ClientBuild|x86
+		ServerBuild|x64 = ServerBuild|x64
+		ServerBuild|x86 = ServerBuild|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientBuild|x64.ActiveCfg = Client|x64
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientBuild|x64.Build.0 = Client|x64
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ClientBuild|x86.ActiveCfg = Server|Win32
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerBuild|x64.ActiveCfg = Server|x64
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerBuild|x64.Build.0 = Server|x64
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerBuild|x86.ActiveCfg = Server|Win32
+		{1E4000B2-AFD7-45F8-B38D-2BF97E7806BF}.ServerBuild|x86.Build.0 = Server|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {35F3DEE3-E7BE-42A2-87EE-7D48440AA441}
+	EndGlobalSection
+EndGlobal
